backend/requirements.txt
Flask
opencv-python
numpy<2.0
Pillow
flask-cors

